+++ CHANGELOG.rst
@@ -1,5 +1,12 @@
 Changelog
 =========
+0.1.4
+-----
+2015-10-10
+
+- Added extra command to reset the settings (clean database, the destination).
+- Now when listing installed version, the active version is indicated.
+
 0.1.3
 -----
 2015-10-10